+++ SpeedyWeather/src/output/feedback.jl
@@ -51,7 +51,9 @@ end
 """
 $(TYPEDSIGNATURES)
 Initializes the a `Feedback` struct."""
-function initialize!(feedback::Feedback, clock::Clock, model::AbstractModel)
+function initialize!(feedback::Feedback, variables::Variables, model::AbstractModel)
+    (; clock) = variables.prognostic
+
     # set to false to recheck for NaNs
     feedback.nans_detected = false
 
@@ -70,7 +72,7 @@ function initialize!(feedback::Feedback, clock::Clock, model::AbstractModel)
     (; showspeed, description, verbose, feedback_dt) = feedback
     desc = description * (model.output.active ? " $(model.output.run_folder) " : " ")
     feedback.progress_meter = ProgressMeter.Progress(
-        clock.n_timesteps - 1;
+        clock.n_steps;          # use time stepper steps (regardless Δt) not time steps of size Δt
         enabled = verbose,
         showspeed,
         desc,
@@ -83,24 +85,21 @@ function initialize!(feedback::Feedback, clock::Clock, model::AbstractModel)
     return nothing
 end
 
-# fallback if feedback is set to nothing
-initialize!(::Nothing, clock::Clock, model::AbstractModel) = nothing
-
 progress!(feedback::Feedback) = ProgressMeter.next!(feedback.progress_meter)
 
-function progress!(feedback::Feedback, vars::Variables)
+function progress!(feedback::Feedback, vars::Variables, model::AbstractModel)
     every_nsteps = feedback.progress_meter.core.check_iterations
     (; counter) = feedback.progress_meter.core
     FEEDBACK_TIME[] = vars.prognostic.clock.time
     feedback.show_umax && mod(counter, every_nsteps) == 0 && max_speed(vars)
     feedback.show_temperature_range && mod(counter, every_nsteps) == 0 && temperature_range(vars)
     progress!(feedback)
-    feedback.debug && nan_detection!(feedback, vars)
+    feedback.debug && nan_detection!(feedback, vars, model)
     return nothing
 end
 
 # fallback for feedback = nothing
-progress!(::Nothing, vars::Variables) = nothing
+progress!(::Nothing, vars::Variables, model::AbstractModel) = nothing
 
 """
 $(TYPEDSIGNATURES)
@@ -112,10 +111,11 @@ finalize!(::Nothing) = nothing
 
 """$(TYPEDSIGNATURES)
 Detect NaN (Not-a-Number, or Inf) in the prognostic variables."""
-function nan_detection!(feedback::Feedback, vars::Variables)
-    feedback.nans_detected && return nothing                        # escape immediately if nans already detected
-    i = feedback.progress_meter.counter                             # time step
-    GPUArrays.@allowscalar vor0 = vars.prognostic.vorticity[2, end, 2]    # only check 1-0 mode of surface vorticity
+function nan_detection!(feedback::Feedback, vars::Variables, model::AbstractModel)
+    feedback.nans_detected && return nothing        # escape immediately if nans already detected
+    i = feedback.progress_meter.counter             # time step
+    vor = get_prognostic_step(vars.prognostic.vorticity, model.time_stepping, feedback)
+    GPUArrays.@allowscalar vor0 = vor[2, end]       # only check 1-0 mode of surface vorticity
 
     # just check first harmonic, spectral transform propagates NaNs globally anyway
     (; time) = vars.prognostic.clock                                # current time for feedback