+++ SpeedyWeather/src/parameterizations/radiation/shortwave_transmissivity.jl
@@ -90,13 +90,14 @@ initialize!(::BackgroundShortwaveTransmissivity, ::AbstractModel) = nothing
     ) = transmissivity
     (; cloud_top, cloud_cover) = clouds
 
-    humid = vars.grid.humidity_prev
+    humid = get_prognostic_step(vars.grid.humidity, model.time_stepping, transmissivity)
     cos_zenith = vars.parameterizations.cos_zenith[ij]
     nlayers = size(t, 2)
 
     sigma_levels = model.geometry.σ_levels_half
     sigma_levels_full = model.geometry.σ_levels_full
-    normalized_surface_pressure = vars.grid.pressure_prev[ij] / 100000
+    pₛ = vars.parameterizations.surface_pressure[ij]          # surface pressure [Pa]
+    normalized_surface_pressure = pₛ / 100000
 
     # Zenith angle correction factor
     azen = transmissivity.zenith_amplitude